refactor(frontend): RouteMarker components
